Allow pluggable index delegate
Move index logic into fakeindex and allow user to specify a different
module to use to handle the index routes.
